+++ client/components/SharedTxnCreateForm.tsx
@@ -2,7 +2,7 @@ import { useUserContext } from 'context/user';
 import { Tracker } from 'pages/shared/trackers';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { useSWRConfig } from 'swr';
-import { plainDateISONowString } from 'utils/temporal';
+import { plainDateISONowString } from 'utils/dates';
 import SharedTxnFormBase, {
   createSharedTxnFormData,
   SharedTxnFormInputs,
